+++ Calculus3/Calculus3/Calculus3.vcxproj
@@ -142,7 +142,11 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Main.cpp" />
+    <ClCompile Include="CalculusSolver.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="CalculusSolver.h" />
+    <ClInclude Include="Functions.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
