+++ Calculus3/Calculus3/Calculus3.vcxproj.filters
@@ -15,7 +15,15 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Main.cpp">
+    <ClInclude Include="CalculusSolver.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Functions.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="CalculusSolver.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
